requirements.txt
lsm-db
zstd
quarry
mmh3
PyYAML

# optional on windows
leveldb
